+++ osrparse/__init__.py
@@ -1,8 +1,9 @@
-from osrparse.replay import (parse_replay_file, parse_replay, ReplayEvent,
-    GameMode, Mod)
+from osrparse.enums import GameMode, Mod
+from osrparse.parse import parse_replay_file, parse_replay
+from osrparse.replay import Replay, ReplayEvent
 
 __version__ = "6.4.1"
 
 
-__all__ = ["parse_replay_file", "parse_replay", "ReplayEvent", "GameMode",
-    "Mod"]
+__all__ = ["GameMode", "Mod", "parse_replay_file", "parse_replay", "Replay",
+    "ReplayEvent"]
